requirements.txt
# Elasticsearch client
elasticsearch>=8.11.0,<9.0.0

# MySQL connector
PyMySQL>=1.1.0

# Configuration management
PyYAML>=6.0
python-dotenv>=1.0.0

# Utilities
python-dateutil>=2.8.2

# Optional: For scheduling
APScheduler>=3.10.0

# Development dependencies (optional)
# pytest>=7.4.0
# pytest-cov>=4.1.0
# black>=23.0.0
# flake8>=6.1.0
# mypy>=1.7.0

